[android][task-manager] Task manager for bare in Expo (#6828)
Task manager for bare android and iOS applications.

+++ packages/expo-sensors/android/src/main/java/expo/modules/sensors/services/SensorServiceSubscription.java
@@ -42,7 +42,6 @@ public class SensorServiceSubscription implements org.unimodules.interfaces.sens
   }
 
   public void stop() {
-    assertSubscriptionIsAlive();
     if (mIsEnabled) {
       mIsEnabled = false;
       mSubscribableSensorService.onSubscriptionEnabledChanged(this);
@@ -50,9 +49,10 @@ public class SensorServiceSubscription implements org.unimodules.interfaces.sens
   }
 
   public void release() {
-    assertSubscriptionIsAlive();
-    mSubscribableSensorService.removeSubscription(this);
-    mHasBeenReleased = true;
+    if (!mHasBeenReleased) {
+      mSubscribableSensorService.removeSubscription(this);
+      mHasBeenReleased = true;
+    }
   }
 
   private void assertSubscriptionIsAlive() {
